Add the new provision exception API.
Guice currently has an Exception called ErrorsException that is used to signal certain kinds of errors.  The general strategy is either to
* call Errors.addError if you come across an error but can keep going
* call throw Errors.addError(...).toException() to report an error when you can't keep going

The second option throws an ErrorsException.  Which is then caught at various api boundaries.  The idea in this cl is to introduce a parallel API called InternalProvisionException which can take over the analogous usecases for provisioning time errors and ErrorsException will be preserved for configuration time errors. ErrorsException will be renamed to InternalConfigurationException after InternalProvisionException is fully integrated.

This also refactors Errors.java to move a bunch of static utilities for dealing with Message objects to a new class Messages

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=177714992
