+++ core/src/com/google/inject/CreationException.java
@@ -19,7 +19,7 @@ package com.google.inject;
 import static com.google.common.base.Preconditions.checkArgument;
 
 import com.google.common.collect.ImmutableSet;
-import com.google.inject.internal.Errors;
+import com.google.inject.internal.Messages;
 import com.google.inject.spi.Message;
 import java.util.Collection;
 
@@ -37,7 +37,7 @@ public class CreationException extends RuntimeException {
   public CreationException(Collection<Message> messages) {
     this.messages = ImmutableSet.copyOf(messages);
     checkArgument(!this.messages.isEmpty());
-    initCause(Errors.getOnlyCause(this.messages));
+    initCause(Messages.getOnlyCause(this.messages));
   }
 
   /** Returns messages for the errors that caused this exception. */
@@ -47,7 +47,7 @@ public class CreationException extends RuntimeException {
 
   @Override
   public String getMessage() {
-    return Errors.format("Unable to create injector, see the following errors", messages);
+    return Messages.format("Unable to create injector, see the following errors", messages);
   }
 
   private static final long serialVersionUID = 0;
