+++ core/src/com/google/inject/internal/Errors.java
@@ -16,15 +16,9 @@
 
 package com.google.inject.internal;
 
-import static com.google.common.base.Preconditions.checkArgument;
-
-import com.google.common.base.Equivalence;
-import com.google.common.base.Objects;
-import com.google.common.base.Throwables;
 import com.google.common.collect.ImmutableList;
 import com.google.common.collect.ImmutableSet;
 import com.google.common.collect.Lists;
-import com.google.common.collect.Maps;
 import com.google.common.collect.Ordering;
 import com.google.common.primitives.Primitives;
 import com.google.inject.Binding;
@@ -38,13 +32,11 @@ import com.google.inject.Provides;
 import com.google.inject.ProvisionException;
 import com.google.inject.Scope;
 import com.google.inject.TypeLiteral;
-import com.google.inject.internal.util.Classes;
 import com.google.inject.internal.util.SourceProvider;
 import com.google.inject.internal.util.StackTraceElements;
 import com.google.inject.spi.Dependency;
 import com.google.inject.spi.ElementSource;
 import com.google.inject.spi.InjectionListener;
-import com.google.inject.spi.InjectionPoint;
 import com.google.inject.spi.Message;
 import com.google.inject.spi.ScopeBinding;
 import com.google.inject.spi.TypeConverterBinding;
@@ -57,7 +49,6 @@ import java.lang.reflect.Member;
 import java.lang.reflect.Method;
 import java.lang.reflect.Type;
 import java.util.ArrayList;
-import java.util.Arrays;
 import java.util.Collection;
 import java.util.Collections;
 import java.util.Formatter;
@@ -229,7 +220,7 @@ public final class Errors implements Serializable {
         String have = bindingKey.getTypeLiteral().toString();
         if (have.contains(want) || want.contains(have)) {
           Formatter fmt = new Formatter();
-          formatSource(fmt, bindingMap.get(bindingKey).getSource());
+          Messages.formatSource(fmt, bindingMap.get(bindingKey).getSource());
           String match = String.format("%s bound%s", convert(bindingKey), fmt.toString());
           possibleMatches.add(match);
           // TODO: Consider a check that if there are more than some number of results,
@@ -262,7 +253,7 @@ public final class Errors implements Serializable {
     return addMessage(sb.toString());
   }
 
-  public Errors jitDisabled(Key key) {
+  public Errors jitDisabled(Key<?> key) {
     return addMessage("Explicit bindings are required and %s is not explicitly bound.", key);
   }
 
@@ -628,28 +619,10 @@ public final class Errors implements Serializable {
     throw new ProvisionException(getMessages());
   }
 
-  private Message merge(Message message) {
-    List<Object> sources = Lists.newArrayList();
-    sources.addAll(getSources());
-    List<Object> messageSources = message.getSources();
-    // It is possible that the end of getSources() and the beginning of message.getSources() are
-    // equivalent, in this case we should drop the repeated source when joining the lists.  The
-    // most likely scenario where this would happen is when a scoped binding throws an exception,
-    // due to the fact that InternalFactoryToProviderAdapter applies the binding source when
-    // merging errors.
-    if (!sources.isEmpty()
-        && !messageSources.isEmpty()
-        && Objects.equal(messageSources.get(0), sources.get(sources.size() - 1))) {
-      messageSources = messageSources.subList(1, messageSources.size());
-    }
-    sources.addAll(messageSources);
-
-    return new Message(sources, message.getMessage(), message.getCause());
-  }
-
   public Errors merge(Collection<Message> messages) {
+    List<Object> sources = getSources();
     for (Message message : messages) {
-      addMessage(merge(message));
+      addMessage(Messages.mergeSources(sources, message));
     }
     return this;
   }
@@ -660,6 +633,11 @@ public final class Errors implements Serializable {
     }
 
     merge(moreErrors.root.errors);
+    return this;
+  }
+
+  public Errors merge(InternalProvisionException ipe) {
+    merge(ipe.getErrors());
     return this;
   }
 
@@ -694,8 +672,7 @@ public final class Errors implements Serializable {
   }
 
   private Errors addMessage(Throwable cause, String messageFormat, Object... arguments) {
-    String message = format(messageFormat, arguments);
-    addMessage(new Message(getSources(), message, cause));
+    addMessage(Messages.create(cause, getSources(), messageFormat, arguments));
     return this;
   }
 
@@ -707,12 +684,14 @@ public final class Errors implements Serializable {
     return this;
   }
 
+  // TODO(lukes): the format(...) overloads violate best practices since they do very different
+  // things. Rename them as part of the inlining process.
+
+  // TODO(lukes): inline into callers
   public static String format(String messageFormat, Object... arguments) {
-    for (int i = 0; i < arguments.length; i++) {
-      arguments[i] = Errors.convert(arguments[i]);
-    }
-    return String.format(messageFormat, arguments);
+    return Messages.format(messageFormat, arguments);
   }
+
 
   public List<Message> getMessages() {
     if (root.errors == null) {
@@ -727,48 +706,13 @@ public final class Errors implements Serializable {
     }.sortedCopy(root.errors);
   }
 
+  // TODO(lukes): remove external callers
   /** Returns the formatted message for an exception with the specified messages. */
   public static String format(String heading, Collection<Message> errorMessages) {
-    Formatter fmt = new Formatter().format(heading).format(":%n%n");
-    int index = 1;
-    boolean displayCauses = getOnlyCause(errorMessages) == null;
-
-    Map<Equivalence.Wrapper<Throwable>, Integer> causes = Maps.newHashMap();
-    for (Message errorMessage : errorMessages) {
-      int thisIdx = index++;
-      fmt.format("%s) %s%n", thisIdx, errorMessage.getMessage());
-
-      List<Object> dependencies = errorMessage.getSources();
-      for (int i = dependencies.size() - 1; i >= 0; i--) {
-        Object source = dependencies.get(i);
-        formatSource(fmt, source);
-      }
-
-      Throwable cause = errorMessage.getCause();
-      if (displayCauses && cause != null) {
-        Equivalence.Wrapper<Throwable> causeEquivalence = ThrowableEquivalence.INSTANCE.wrap(cause);
-        if (!causes.containsKey(causeEquivalence)) {
-          causes.put(causeEquivalence, thisIdx);
-          fmt.format("Caused by: %s", Throwables.getStackTraceAsString(cause));
-        } else {
-          int causeIdx = causes.get(causeEquivalence);
-          fmt.format(
-              "Caused by: %s (same stack trace as error #%s)",
-              cause.getClass().getName(), causeIdx);
-        }
-      }
-
-      fmt.format("%n");
-    }
-
-    if (errorMessages.size() == 1) {
-      fmt.format("1 error");
-    } else {
-      fmt.format("%s errors", errorMessages.size());
-    }
-
-    return fmt.toString();
+    return Messages.format(heading, errorMessages);
   }
+
+  // TODO(lukes): move @Provides logic into @Provides methods to help simplify this monster
 
   /**
    * Returns {@code value} if it is non-null or allowed to be null. Otherwise a message is added and
@@ -776,6 +720,8 @@ public final class Errors implements Serializable {
    */
   public <T> T checkForNull(T value, Object source, Dependency<?> dependency)
       throws ErrorsException {
+    // TODO(lukes): move this method to InternalProvisionException?  Maybe split into 2, one for
+    // ProviderMethods and one for everything else.
     if (value != null || dependency.isNullable()) {
       return value;
     }
@@ -791,23 +737,24 @@ public final class Errors implements Serializable {
             return value; // user doesn't care about injecting nulls to non-@Nullables.
           case WARN:
             // Warn only once, otherwise we spam logs too much.
-            if (!warnedDependencies.add(dependency)) {
-              return value;
+            if (warnedDependencies.add(dependency)) {
+              logger.log(
+                  Level.WARNING,
+                  "Guice injected null into {0} (a {1}), please mark it @Nullable."
+                      + " Use -Dguice_check_nullable_provides_params=ERROR to turn this into an"
+                      + " error.",
+                  new Object[] {
+                    Messages.formatParameter(dependency), Messages.convert(dependency.getKey())
+                  });
             }
-            logger.log(
-                Level.WARNING,
-                "Guice injected null into {0} (a {1}), please mark it @Nullable."
-                    + " Use -Dguice_check_nullable_provides_params=ERROR to turn this into an"
-                    + " error.",
-                new Object[] {formatParameter(dependency), convert(dependency.getKey())});
-            return null; // log & exit.
+            return value;
         }
       }
     }
 
     Object formattedDependency =
         (dependency.getParameterIndex() != -1)
-            ? formatParameter(dependency)
+            ? Messages.formatParameter(dependency)
             : StackTraceElements.forMember(dependency.getInjectionPoint().getMember());
     addMessage(
         "null returned by binding at %s%n but %s is not @Nullable", source, formattedDependency);
@@ -815,253 +762,23 @@ public final class Errors implements Serializable {
     throw toException();
   }
 
-  /**
-   * Returns the cause throwable if there is exactly one cause in {@code messages}. If there are
-   * zero or multiple messages with causes, null is returned.
-   */
-  public static Throwable getOnlyCause(Collection<Message> messages) {
-    Throwable onlyCause = null;
-    for (Message message : messages) {
-      Throwable messageCause = message.getCause();
-      if (messageCause == null) {
-        continue;
-      }
-
-      if (onlyCause != null && !ThrowableEquivalence.INSTANCE.equivalent(onlyCause, messageCause)) {
-        return null;
-      }
-
-      onlyCause = messageCause;
-    }
-
-    return onlyCause;
-  }
-
   public int size() {
     return root.errors == null ? 0 : root.errors.size();
   }
 
-  private abstract static class Converter<T> {
-
-    final Class<T> type;
-
-    Converter(Class<T> type) {
-      this.type = type;
-    }
-
-    boolean appliesTo(Object o) {
-      return o != null && type.isAssignableFrom(o.getClass());
-    }
-
-    String convert(Object o) {
-      return toString(type.cast(o));
-    }
-
-    abstract String toString(T t);
-  }
-
-  private static final Collection<Converter<?>> converters =
-      ImmutableList.of(
-          new Converter<Class>(Class.class) {
-            @Override
-            public String toString(Class c) {
-              return c.getName();
-            }
-          },
-          new Converter<Member>(Member.class) {
-            @Override
-            public String toString(Member member) {
-              return Classes.toString(member);
-            }
-          },
-          new Converter<Key>(Key.class) {
-            @Override
-            public String toString(Key key) {
-              if (key.getAnnotationType() != null) {
-                return key.getTypeLiteral()
-                    + " annotated with "
-                    + (key.getAnnotation() != null ? key.getAnnotation() : key.getAnnotationType());
-              } else {
-                return key.getTypeLiteral().toString();
-              }
-            }
-          });
-
+  // TODO(lukes): inline in callers.  There are some callers outside of guice, so this is difficult
   public static Object convert(Object o) {
-    ElementSource source = null;
-    if (o instanceof ElementSource) {
-      source = (ElementSource) o;
-      o = source.getDeclaringSource();
-    }
-    return convert(o, source);
+    return Messages.convert(o);
   }
 
+  // TODO(lukes): inline in callers.  There are some callers outside of guice, so this is difficult
   public static Object convert(Object o, ElementSource source) {
-    for (Converter<?> converter : converters) {
-      if (converter.appliesTo(o)) {
-        return appendModules(converter.convert(o), source);
-      }
-    }
-    return appendModules(o, source);
+    return Messages.convert(o, source);
   }
 
-  private static Object appendModules(Object source, ElementSource elementSource) {
-    String modules = moduleSourceString(elementSource);
-    if (modules.length() == 0) {
-      return source;
-    } else {
-      return source + modules;
-    }
-  }
-
-  private static String moduleSourceString(ElementSource elementSource) {
-    // if we only have one module (or don't know what they are), then don't bother
-    // reporting it, because the source already is going to report exactly that module.
-    if (elementSource == null) {
-      return "";
-    }
-    List<String> modules = Lists.newArrayList(elementSource.getModuleClassNames());
-    // Insert any original element sources w/ module info into the path.
-    while (elementSource.getOriginalElementSource() != null) {
-      elementSource = elementSource.getOriginalElementSource();
-      modules.addAll(0, elementSource.getModuleClassNames());
-    }
-    if (modules.size() <= 1) {
-      return "";
-    }
-
-    // Ideally we'd do:
-    //    return Joiner.on(" -> ")
-    //        .appendTo(new StringBuilder(" (via modules: "), Lists.reverse(modules))
-    //        .append(")").toString();
-    // ... but for some reason we can't find Lists.reverse, so do it the boring way.
-    StringBuilder builder = new StringBuilder(" (via modules: ");
-    for (int i = modules.size() - 1; i >= 0; i--) {
-      builder.append(modules.get(i));
-      if (i != 0) {
-        builder.append(" -> ");
-      }
-    }
-    builder.append(")");
-    return builder.toString();
-  }
-
+  // TODO(lukes): inline in callers.  There are some callers outside of guice, so this is difficult
   public static void formatSource(Formatter formatter, Object source) {
-    ElementSource elementSource = null;
-    if (source instanceof ElementSource) {
-      elementSource = (ElementSource) source;
-      source = elementSource.getDeclaringSource();
-    }
-    formatSource(formatter, source, elementSource);
+    Messages.formatSource(formatter, source);
   }
 
-  public static void formatSource(Formatter formatter, Object source, ElementSource elementSource) {
-    String modules = moduleSourceString(elementSource);
-    if (source instanceof Dependency) {
-      Dependency<?> dependency = (Dependency<?>) source;
-      InjectionPoint injectionPoint = dependency.getInjectionPoint();
-      if (injectionPoint != null) {
-        formatInjectionPoint(formatter, dependency, injectionPoint, elementSource);
-      } else {
-        formatSource(formatter, dependency.getKey(), elementSource);
-      }
-
-    } else if (source instanceof InjectionPoint) {
-      formatInjectionPoint(formatter, null, (InjectionPoint) source, elementSource);
-
-    } else if (source instanceof Class) {
-      formatter.format("  at %s%s%n", StackTraceElements.forType((Class<?>) source), modules);
-
-    } else if (source instanceof Member) {
-      formatter.format("  at %s%s%n", StackTraceElements.forMember((Member) source), modules);
-
-    } else if (source instanceof TypeLiteral) {
-      formatter.format("  while locating %s%s%n", source, modules);
-
-    } else if (source instanceof Key) {
-      Key<?> key = (Key<?>) source;
-      formatter.format("  while locating %s%n", convert(key, elementSource));
-
-    } else if (source instanceof Thread) {
-      formatter.format("  in thread %s%n", source);
-
-    } else {
-      formatter.format("  at %s%s%n", source, modules);
-    }
-  }
-
-  public static void formatInjectionPoint(
-      Formatter formatter,
-      Dependency<?> dependency,
-      InjectionPoint injectionPoint,
-      ElementSource elementSource) {
-    Member member = injectionPoint.getMember();
-    Class<? extends Member> memberType = Classes.memberType(member);
-
-    if (memberType == Field.class) {
-      dependency = injectionPoint.getDependencies().get(0);
-      formatter.format("  while locating %s%n", convert(dependency.getKey(), elementSource));
-      formatter.format("    for field at %s%n", StackTraceElements.forMember(member));
-
-    } else if (dependency != null) {
-      formatter.format("  while locating %s%n", convert(dependency.getKey(), elementSource));
-      formatter.format("    for %s%n", formatParameter(dependency));
-
-    } else {
-      formatSource(formatter, injectionPoint.getMember());
-    }
-  }
-
-  private static String formatParameter(Dependency<?> dependency) {
-    int ordinal = dependency.getParameterIndex() + 1;
-    return String.format(
-        "the %s%s parameter of %s",
-        ordinal,
-        getOrdinalSuffix(ordinal),
-        StackTraceElements.forMember(dependency.getInjectionPoint().getMember()));
-  }
-
-  /**
-   * Maps {@code 1} to the string {@code "1st"} ditto for all non-negative numbers
-   *
-   * @see <a href="https://en.wikipedia.org/wiki/English_numerals#Ordinal_numbers">
-   *     https://en.wikipedia.org/wiki/English_numerals#Ordinal_numbers</a>
-   */
-  private static String getOrdinalSuffix(int ordinal) {
-    // negative ordinals don't make sense, we allow zero though because we are programmers
-    checkArgument(ordinal >= 0);
-    if ((ordinal / 10) % 10 == 1) {
-      // all the 'teens' are weird
-      return "th";
-    } else {
-      // could use a lookup table? any better?
-      switch (ordinal % 10) {
-        case 1:
-          return "st";
-        case 2:
-          return "nd";
-        case 3:
-          return "rd";
-        default:
-          return "th";
-      }
-    }
-  }
-
-  static class ThrowableEquivalence extends Equivalence<Throwable> {
-    static final ThrowableEquivalence INSTANCE = new ThrowableEquivalence();
-
-    @Override
-    protected boolean doEquivalent(Throwable a, Throwable b) {
-      return a.getClass().equals(b.getClass())
-          && Objects.equal(a.getMessage(), b.getMessage())
-          && Arrays.equals(a.getStackTrace(), b.getStackTrace())
-          && equivalent(a.getCause(), b.getCause());
-    }
-
-    @Override
-    protected int doHash(Throwable t) {
-      return Objects.hashCode(t.getClass().hashCode(), t.getMessage(), hash(t.getCause()));
-    }
-  }
 }
